test(api): migrate realtime token to api-first bdd
Round 25 of the API-first BDD migration (#16967), REALTIME TOKEN.

- Extend createBddApi with the realtimeToken client.
- Add zero-realtime-token.bdd.test.ts: mint a token for an authenticated caller
  and assert (via the Ably mock, the only observable surface) the subscribe-only
  capability scoped to the caller's own user channel with the hour TTL; an
  unauthenticated request is 401 and mints nothing.
- Delete zero-realtime-token.test.ts whole.

Coverage verified (source-only): zero-realtime-token.ts (11/2/1) and realtime.ts
(38/9/16) unchanged vs main. No regressions.
